Day 14 part 2! that was a dumb bug

+++ Day14/Day14.h
@@ -9,7 +9,7 @@ class Day14 {
 public:
     static int Part1();
 
-    static int Part2();
+    static long long Part2();
 };
 
 #endif //DAY14_H
